backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0

